requirements.txt
streamlit>=1.44.0
crewai>=0.114.0
crewai-tools>=0.40.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.3.0
langchain-google-genai>=2.1.0
langchain-text-splitters>=0.3.0
matplotlib>=3.8.0
pypdf>=5.4.0
python-dotenv>=1.0.1
reportlab>=4.2.0
litellm>=1.63.0
google-genai
requests
